+++ lib/src/bottom_sheet/theme/default_sbb_bottom_sheet_theme_data.dart
@@ -6,7 +6,7 @@ class DefaultSBBBottomSheetThemeData extends SBBBottomSheetThemeData {
   DefaultSBBBottomSheetThemeData(SBBBaseStyle baseStyle)
     : super(
         style: SBBBottomSheetStyle(
-          titleTextStyle: baseStyle.themedTextStyle(textStyle: baseStyle.textTheme.largeLight),
+          titleTextStyle: baseStyle.textTheme.largeLight,
           titleForegroundColor: baseStyle.colorScheme.textPrimary,
           leadingTextStyle: baseStyle.textTheme.defaultTextStyle,
           leadingForegroundColor: baseStyle.colorScheme.iconPrimary,
@@ -18,7 +18,7 @@ class DefaultSBBBottomSheetThemeData extends SBBBottomSheetThemeData {
           barrierColor: const Color(0x80000000),
           titleBodyGap: SBBSpacing.small,
           titleMinHeight: SBBSpacing.xLarge,
-          padding: .symmetric(horizontal: SBBSpacing.medium).copyWith(top: SBBSpacing.small),
+          padding: .symmetric(horizontal: SBBSpacing.medium, vertical: SBBSpacing.small),
         ),
       );
 }