+++ lib/src/decorated_text/sbb_decorated_text.dart
@@ -32,8 +32,6 @@ import 'package:sbb_design_system_mobile/src/shared/debug.dart';
 ///
 /// * [value]: The static text to display (cannot be edited)
 /// * [onTap]: Called when the widget is tapped. Provides visual feedback via InkWell
-/// * [enabled]: Controls interactivity. When false, taps are ignored and disabled
-///   styling is applied
 /// * [decoration]: Customizes the decoration surrounding the value, including icons,
 ///   labels, and error states
 /// * [maxLines], [minLines], [expands]: Control text layout similar to [SBBTextInput]
@@ -107,8 +105,7 @@ class SBBDecoratedText extends StatefulWidget {
 
   /// Called when the widget is tapped.
   ///
-  /// The tap triggers [InkWell] visual feedback (ripple/highlight). This callback
-  /// is only invoked if [enabled] is true.
+  /// The tap triggers [InkWell] visual feedback (ripple/highlight).
   final GestureTapCallback? onTap;
 
   /// Customizes the visual appearance of the decorated text.
@@ -264,6 +261,7 @@ class SBBDecoratedTextBoxed extends SBBDecoratedText {
     super.minLines,
     super.expands,
     super.style,
+    this.margin,
   }) : super(
          decoration: decoration?.contentPadding != null
              ? decoration
@@ -271,6 +269,9 @@ class SBBDecoratedTextBoxed extends SBBDecoratedText {
                  contentPadding: .symmetric(horizontal: SBBSpacing.medium),
                ),
        );
+
+  /// The margin of the content box surrounding the [SBBDecoratedText].
+  final EdgeInsetsGeometry? margin;
 
   @override
   State<SBBDecoratedText> createState() => _SBBTextInputStateBoxed();
@@ -282,6 +283,6 @@ class _SBBTextInputStateBoxed extends _SBBDecoratedTextState {
 
   @override
   Widget build(BuildContext context) {
-    return SBBContentBox(child: super.build(context));
+    return SBBContentBox(margin: (widget as SBBDecoratedTextBoxed).margin, child: super.build(context));
   }
 }