+++ lib/src/header_box/header_box_content.dart
@@ -1,5 +1,9 @@
+import 'dart:math' as math;
+
 import 'package:flutter/material.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
+import 'package:sbb_design_system_mobile/src/header_box/utils.dart';
+import 'package:sbb_design_system_mobile/src/shared/utils.dart';
 
 class HeaderBoxContent extends StatelessWidget {
   const HeaderBoxContent({
@@ -23,21 +27,23 @@ class HeaderBoxContent extends StatelessWidget {
   final Widget? trailing;
   final SBBHeaderBoxStyle style;
 
+  static const double _textPadding = 2.0;
+
   @override
   Widget build(BuildContext context) {
-    final titleWidget = _addDefaultAncestorWithResolved(
+    final titleWidget = addDefaultAncestorWithResolved(
       textStyle: style.titleTextStyle,
       foregroundColor: style.titleForegroundColor,
       child: _resolveTitle(),
     );
 
-    final subtitleWidget = _addDefaultAncestorWithResolved(
+    final subtitleWidget = addDefaultAncestorWithResolved(
       textStyle: style.subtitleTextStyle,
       foregroundColor: style.subtitleForegroundColor,
       child: _resolveSubtitle(),
     );
 
-    final leadingWidget = _addDefaultAncestorWithResolved(
+    final leadingWidget = addDefaultAncestorWithResolved(
       textStyle: style.leadingTextStyle,
       foregroundColor: style.leadingForegroundColor,
       child: _resolveLeading(),
@@ -48,7 +54,7 @@ class HeaderBoxContent extends StatelessWidget {
     }
 
     Widget child = Padding(
-      padding: const EdgeInsets.symmetric(vertical: 2.0),
+      padding: const EdgeInsets.symmetric(vertical: _textPadding),
       child: titleWidget,
     );
     if (leadingWidget != null) {
@@ -56,7 +62,7 @@ class HeaderBoxContent extends StatelessWidget {
         spacing: SBBSpacing.xSmall,
         children: [
           leadingWidget,
-          child,
+          Expanded(child: child),
         ],
       );
     }
@@ -68,7 +74,7 @@ class HeaderBoxContent extends StatelessWidget {
         children: [
           child,
           Padding(
-            padding: const EdgeInsets.symmetric(vertical: 2.0),
+            padding: const EdgeInsets.symmetric(vertical: _textPadding),
             child: subtitleWidget,
           ),
         ],
@@ -115,6 +121,35 @@ class HeaderBoxContent extends StatelessWidget {
 
     return subtitle;
   }
+
+  static double calculateHeight({
+    required SBBHeaderBoxStyle style,
+    PreferredSizeWidget? title,
+    String? titleText,
+    PreferredSizeWidget? subtitle,
+    String? subtitleText,
+    PreferredSizeWidget? leading,
+    IconData? leadingIconData,
+    TextScaler textScaler = TextScaler.noScaling,
+  }) {
+    if (title == null && titleText == null) {
+      return 0.0;
+    }
+
+    final iconSize = IconThemeData.fallback().size!;
+    final titleHeight =
+        calculateHeightWithTextFallback(title, titleText, style.titleTextStyle, textScaler) + _textPadding * 2;
+    final leadingHeight = leading?.preferredSize.height ?? (leadingIconData != null ? iconSize : 0.0);
+    final subtitleHeight = calculateHeightWithTextFallback(subtitle, subtitleText, style.subtitleTextStyle, textScaler);
+
+    final titleRowHeight = math.max(titleHeight, leadingHeight);
+
+    if (subtitleHeight > 0) {
+      return titleRowHeight + style.titleSubtitleGap! + subtitleHeight + _textPadding * 2;
+    } else {
+      return titleRowHeight;
+    }
+  }
 }
 
 class LargeHeaderBoxContent extends StatelessWidget {
@@ -141,18 +176,18 @@ class LargeHeaderBoxContent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final titleWidget = _addDefaultAncestorWithResolved(
+    final titleWidget = addDefaultAncestorWithResolved(
       textStyle: style.titleTextStyle,
       foregroundColor: style.titleForegroundColor,
       child: _resolveTitle(),
     );
-    final subtitleWidget = _addDefaultAncestorWithResolved(
+    final subtitleWidget = addDefaultAncestorWithResolved(
       textStyle: style.subtitleTextStyle,
       foregroundColor: style.subtitleForegroundColor,
       child: _resolveSubtitle(),
     );
 
-    final leadingWidget = _addDefaultAncestorWithResolved(
+    final leadingWidget = addDefaultAncestorWithResolved(
       textStyle: style.leadingTextStyle,
       foregroundColor: style.leadingForegroundColor,
       child: _resolveLeading(),
@@ -211,25 +246,4 @@ class LargeHeaderBoxContent extends StatelessWidget {
 
     return subtitle;
   }
-}
-
-Widget? _addDefaultAncestorWithResolved({
-  Widget? child,
-  Color? foregroundColor,
-  TextStyle? textStyle,
-}) {
-  if (child == null) {
-    return null;
-  }
-
-  final resolvedTextStyle = textStyle?.copyWith(color: foregroundColor) ?? TextStyle(color: foregroundColor);
-
-  child = DefaultTextStyle.merge(
-    style: resolvedTextStyle,
-    child: IconTheme.merge(
-      data: IconThemeData(color: foregroundColor),
-      child: child,
-    ),
-  );
-  return child;
 }