+++ lib/src/header_box/theme/default_sbb_header_box_theme_data.dart
@@ -14,8 +14,13 @@ class DefaultSBBHeaderBoxThemeData extends SBBHeaderBoxThemeData {
           labelForegroundColor: baseStyle.colorScheme.textPrimary,
           leadingForegroundColor: baseStyle.colorScheme.iconPrimary,
           trailingForegroundColor: baseStyle.colorScheme.iconPrimary,
-          labelTextStyle: baseStyle.themedTextStyle(textStyle: SBBTextStyles.smallLight),
+          labelTextStyle: baseStyle.textTheme.smallLight,
           padding: EdgeInsets.fromLTRB(SBBSpacing.medium, SBBSpacing.xSmall, SBBSpacing.medium, SBBSpacing.xSmall),
         ),
       );
+
+  /// Fallback sbb header box theme used to calculate the preferred size without a build context.
+  static final fallback = DefaultSBBHeaderBoxThemeData(
+    baseStyle: SBBBaseStyle.$default(brightness: Brightness.light, themeContext: .sbb),
+  );
 }