+++ lib/src/header_box/theme/sbb_header_box_style.dart
@@ -44,11 +44,9 @@ class SBBHeaderBoxStyle {
 
   factory SBBHeaderBoxStyle.$default({required SBBBaseStyle baseStyle}) {
     return SBBHeaderBoxStyle(
-      titleTextStyle: baseStyle.themedTextStyle(
-        textStyle: SBBTextStyles.mediumBold.copyWith(overflow: TextOverflow.ellipsis),
-      ),
+      titleTextStyle: baseStyle.textTheme.mediumBold?.copyWith(overflow: TextOverflow.ellipsis),
       titleForegroundColor: baseStyle.colorScheme.textPrimary,
-      subtitleTextStyle: baseStyle.themedTextStyle(textStyle: SBBTextStyles.smallLight),
+      subtitleTextStyle: baseStyle.textTheme.smallLight,
       subtitleForegroundColor: baseStyle.colorScheme.textSecondary,
       leadingForegroundColor: baseStyle.colorScheme.iconPrimary,
       backgroundColor: baseStyle.colorScheme.backgroundContent,
@@ -158,7 +156,7 @@ class SBBHeaderBoxStyle {
   static const BorderRadius radius = BorderRadius.all(Radius.circular(SBBSpacing.medium));
 
   /// The minimum height of the header box without external margin applied.
-  static const minHeight = 56.0;
+  static const double minHeight = 56.0;
 
   SBBHeaderBoxStyle copyWith({
     TextStyle? titleTextStyle,