+++ lib/src/segmented_button/sbb_segmented_button.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:sbb_design_system_mobile/src/segmented_button/theme/default_sbb_segmented_button_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/shared/debug.dart';
+import 'package:sbb_design_system_mobile/src/shared/utils.dart';
 
 /// The SBB Segmented Button.
 ///
@@ -212,13 +213,13 @@ class _SBBSegmentedButtonState<T> extends State<SBBSegmentedButton<T>> {
     final themeData = Theme.of(context).sbbSegmentedButtonTheme;
     final effectiveLeadingGapWidth = widget.leadingHorizontalGapWidth ?? themeData.leadingHorizontalGapWidth!;
 
-    leading = _addDefaultAncestorWithResolved(
+    leading = addDefaultAncestorWithResolved(
       child: leading,
       foregroundColor: foregroundColor,
       textStyle: resolvedTextStyle,
     );
 
-    label = _addDefaultAncestorWithResolved(
+    label = addDefaultAncestorWithResolved(
       child: label,
       foregroundColor: foregroundColor,
       textStyle: resolvedTextStyle,
@@ -240,25 +241,6 @@ class _SBBSegmentedButtonState<T> extends State<SBBSegmentedButton<T>> {
 
     return child;
   }
-
-  Widget? _addDefaultAncestorWithResolved({
-    Widget? child,
-    required Color? foregroundColor,
-    TextStyle? textStyle,
-  }) {
-    if (child == null) return null;
-
-    final resolvedTextStyle = textStyle?.copyWith(color: foregroundColor) ?? TextStyle(color: foregroundColor);
-
-    child = DefaultTextStyle.merge(
-      style: resolvedTextStyle,
-      child: IconTheme.merge(
-        data: IconThemeData(color: foregroundColor),
-        child: child,
-      ),
-    );
-    return child;
-  }
 }
 
 /// Creates the filled style variant of [SBBSegmentedButton].
@@ -278,10 +260,10 @@ class SBBSegmentedButtonFilled<T> extends SBBSegmentedButton<T> {
   });
 
   @override
-  State<SBBSegmentedButton<T>> createState() => _SBBSegmentedButtonStateBoxed();
+  State<SBBSegmentedButton<T>> createState() => _SBBSegmentedButtonStateFilled();
 }
 
-class _SBBSegmentedButtonStateBoxed<T> extends _SBBSegmentedButtonState<T> {
+class _SBBSegmentedButtonStateFilled<T> extends _SBBSegmentedButtonState<T> {
   @override
   SBBSegmentedButtonStyle get effectiveStyle {
     final themeData = Theme.of(context).sbbSegmentedButtonTheme;