+++ CHANGELOG.md
@@ -76,6 +76,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - added `SBBTextInput` as replacement for `SBBTextField` - see [v5 migration guide]
 - added `SBBColorScheme` - see [v5 migration guide]
 - added `SBBThemeContext` - see [v5 migration guide]
+- added `SBBHeaderBoxPreferredSize` and `SBBHeaderBoxFlapPreferredSize` for use in headers
 
 ### Changed
 
@@ -124,6 +125,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - Removed `CascadeColumnParentData` as it is not directly used by clients
 - Removed `SBBTextField` - replaced by `SBBTextInput`
 - Removed `merge` extension on `SBBBaseStyle`
+- Removed `themedTextStyle` method from `SBBBaseStyle`
 - Removed `SBBColors.redDark`, use `SBBColors.red85` instead
 
 ## [4.12.0] - 2026-04-28